+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 Major changes.
 
+- Remove dependency on delta-crdts
 - Remove dependency on uuids
 - Pres can now present a document with a custom root node